fix: post가 최신순으로 정렬이 되지 않던 문제 해결
